pyproject.toml
[tool.poetry]
name = "test_bad_file_descriptor"
version = "0.1.0"
description = ""
authors = ["Mathieu Tarral <mathieu.tarral@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-libcloud = "^3.3.1"
fasteners = "^0.16.3"

[tool.poetry.dev-dependencies]
pyfakefs = "^4.5.1"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
